Refactor data and types structure: moved data files to a separate folder and centralize RoomData type in utils
