+++ src/data/data.ts
@@ -1,0 +1,136 @@
+import { RoomData } from "@/utils/types";
+
+export const RoomDataList: RoomData[] = [
+  {
+    category: "Standard",
+    name: "Studio Room (Seaview)",
+    description: "Studio room description.",
+    image: "/images/rooms/standard/StudioRoom.jpg",
+    price: "₱ 6,800",
+    bed: "2 Queen bed",
+    capacity: "4-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "17 m²"
+  },
+  {
+    category: "Standard",
+    name: "Double Room",
+    description: "Double room description.",
+    image: "/images/rooms/standard/DoubleRoom.jpg",
+    price: "₱ 3,630",
+    bed: "1 Double bed",
+    capacity: "2-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "12 m²"
+  },
+  {
+    category: "Standard",
+    name: "Standard Double Room",
+    description: "Standard Double room description.",
+    image: "/images/rooms/standard/StandardDouble.jpg",
+    price: "₱ 3,300",
+    bed: "1 Double bed",
+    capacity: "2-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "13 m²"
+  },
+   {
+    category: "Dormitory",
+    name: "Dormitory (10pax)",
+    description: "Dormitory (10pax) description.",
+    image: "/images/rooms/dormitory/Dormitory(10pax).jpg",
+    price: "₱ 14,850",
+    bed: "5 bunk beds",
+    capacity: "10-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "44 m²"
+  },
+  {
+    category: "Dormitory",
+    name: "Dormitory (6pax)",
+    description: "Dormitory (6pax) description.",
+    image: "/images/rooms/dormitory/Dormitory(6pax)2.jpg",
+    price: "₱ 9,020",
+    bed: "3 bunk beds",
+    capacity: "6-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "26 m²"
+  },
+  {
+    category: "Dormitory",
+    name: "Bahay Kubo",
+    description: "Bahay Kubo description.",
+    image: "/images/rooms/dormitory/BahayKubo.jpg",
+    price: "₱ 8,250",
+    bed: "2 bunk beds",
+    capacity: "8-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "23 m²"
+  },
+  {
+    category: "Deluxe",
+    name: "Deluxe Studio Room (Seaview)",
+    description: "Deluxe Studio Room (Seaview) room description.",
+    image: "/images/rooms/deluxe/DeluxeStudioRoomSeaview.jpg",
+    price: "₱ 7,500",
+    bed: "2 Double bed",
+    capacity: "4-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "31 m²"
+  },
+  {
+    category: "Deluxe",
+    name: "Deluxe Studio Room",
+    description: "Deluxe Studio Room description.",
+    image: "/images/rooms/deluxe/DeluxeStudioRoom.jpg",
+    price: "₱ 6,380",
+    bed: "2 Double bed",
+    capacity: "4-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "17 m²"
+  },
+  {
+    category: "Deluxe",
+    name: "Deluxe Triple Room",
+    description: "Deluxe Triple room description.",
+    image: "/images/rooms/deluxe/DeluxeTripleRoom.jpg",
+    price: "₱ 4,950",
+    bed: "1 single bed, 1 double bed",
+    capacity: "3-person capacity",
+    breakfast: "No Breakfast included",
+    roomsize: "16 m²"
+  },
+  {
+    category: "Deluxe",
+    name: "Deluxe Double Room (Seaview)",
+    description: "Deluxe Double Room (Seaview) description.",
+    image: "/images/rooms/deluxe/DeluxeDoubleRoomSeaview.jpg",
+    price: "₱ 4,400",
+    bed: "1 double bed",
+    capacity: "2-person capacity",
+    breakfast: "Breakfast included",
+    roomsize: "19 m²"
+  },
+  {
+    category: "Deluxe",
+    name: "Deluxe Double Room",
+    description: "Deluxe Double Room description.",
+    image: "/images/rooms/deluxe/DeluxeDoubleRoom.jpg",
+    price: "₱ 3,850",
+    bed: "1 queen bed",
+    capacity: "2-person capacity",
+    breakfast: "No Breakfast included",
+    roomsize: "15 m²"
+  },
+   {
+    category: "Cabana",
+    name: "Cabana",
+    description: "A cozy beachfront cabana with stunning ocean views and private porch.",
+    image: "/images/rooms/cabana.png",
+    price: "₱ 1,000",
+    bed: "No bed",
+    capacity: "10-person capacity",
+    breakfast: "No breakfast included",
+    roomsize: ""
+  },
+]
